requirements.txt
python-telegram-bot>=21.0.1
openai==1.12.0
google-generativeai>=0.7.2
pillow>=10.3.0
numpy==1.26.2
matplotlib==3.8.2
python-dotenv>=1.0.1
requests==2.31.0
jsonschema==4.20.0
pytest==7.4.3
pytest-asyncio==0.23.2
aiohttp==3.9.1
seaborn==0.13.0
pydantic==2.5.2
tenacity==8.2.3
colorlog==6.7.0
firebase-admin==6.3.0
gunicorn==21.2.0 # For Railway deployment 